FlexSerial write - turn off debug output, enable all Uarts...
Now that it looks like an issue was solved with the multiple Flex Serial  objects.

So turned off debug output.  Which then everything failed.  Found that one set of outputs, was the cause, which looks like timing issue.  Appears to work now with delay(1).  Will experiment more later.  But since this is only at the begin method, not a big deal.

I also updated the many stream output case to uncomment to use more of the flex serial objects