Implement FlexSerial::flush()
The siimple version that simply waits for us to output the last byte in the queue into the shift buffer is not sufficient.  It would end up returning from flush() with part of two characters left to be output.

So this version enables the Timer interrupt to happen when we sense that our queue is empty, and then we need to process two of these timer interrupts.  On the second one we disable it and set the flag that says that we are no longer transmitting.   If we do another write after we start up the trigger, we reset the timer status and again turn off the ISR...

I added a one second timeout into flush() in case it totally hangs.