Pin->FlexIO pin info works for all Teensy 4.x boards
I built tables for all of the possible FlexIO pins on the IMXRT 1060 boards
and then did the mapping indirect from the Teensy pins.

That is I used the Teensy boards Pin table, mux pointer data as
what to match in the FlexIO table. This way should not have to change
the code for any new Teensy 4.x or any variant, as long as they have
any pins defined in a pin table.  Alternatively can add in where you
could pass in a MUX pin register...